Косяков по прежнему дофига, тесты уже запускаются сразу, единственное я у себя чистил миграции и базу
