+++ api_yamdb/api/urls.py
@@ -1,14 +1,19 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from .views import CategoryViewSet, GenreViewSet, TitleViewSet
+from .views import (
+    CategoryViewSet, GenreViewSet, TitleViewSet,
+    UserViewSet, get_jwt_token, register
+)
 
 router_v1 = DefaultRouter()
 router_v1.register('titles', TitleViewSet, basename='title')
 router_v1.register('genres', GenreViewSet, basename='genres')
 router_v1.register('categories', CategoryViewSet, basename='categories')
-
+router_v1.register('users', UserViewSet, basename='users')
 
 urlpatterns = [
     path('v1/', include(router_v1.urls)),
+    path('v1/auth/signup/', register, name='register'),
+    path('v1/auth/token/', get_jwt_token, name='token')
 ]
