+++ api_yamdb/api/views.py
@@ -1,14 +1,25 @@
-from rest_framework import filters, viewsets
+from rest_framework import filters, viewsets, permissions, status
 from django.shortcuts import get_object_or_404
 from reviews.models import Review
 from rest_framework.permissions import AllowAny
+from rest_framework.pagination import PageNumberPagination
 from django_filters.rest_framework import DjangoFilterBackend
+from django.contrib.auth.tokens import default_token_generator
+from django.core.mail import send_mail
+from rest_framework_simplejwt.tokens import AccessToken
+from rest_framework.response import Response
+from rest_framework.decorators import action, api_view, permission_classes
 
 from titles.models import Category, Genre, Title
 from .filters import TitleFilter
+from users.models import User
 from users.permissions import IsModerOrAdminOrOwnerOrReadOnly, IsAdmin
 from .mixins import GetListCreateDeleteMixin
-from .serializers import (CategorySerializer, CommentSerializer, GenreSerializer, ReviewSerializer, TitleSerializer)
+from .serializers import (
+    CategorySerializer, CommentSerializer, GenreSerializer,
+    ReviewSerializer, TitleSerializer, TokenSerializer,
+    RegisterSerializer, TokenSerializer, UserEditSerializer, UserSerializer
+)
 
 
 class ReviewViewSet(viewsets.ModelViewSet):
@@ -32,7 +43,7 @@ class CommentViewSet(viewsets.ModelViewSet):
     def get_queryset(self):
         review = get_object_or_404(Review, id=self.kwargs.get('review_id'))
         return review.comments.all()
-    
+
     def perform_create(self, serializer):
         review = get_object_or_404(Review, id=self.kwargs.get('review_id'))
         serializer.save(author=self.request.user, review=review)
@@ -41,7 +52,7 @@ class CommentViewSet(viewsets.ModelViewSet):
 class TitleViewSet(viewsets.ModelViewSet):
     queryset = Title.objects.all()
     serializer_class = TitleSerializer
-    permission_classes = [IsAdmin, AllowAny,]
+    permission_classes = [IsAdmin, AllowAny, ]
     filter_backends = (DjangoFilterBackend, filters.OrderingFilter,)
     filter_class = TitleFilter
 
@@ -53,7 +64,7 @@ class TitleViewSet(viewsets.ModelViewSet):
 class CategoryViewSet(GetListCreateDeleteMixin):
     queryset = Category.objects.all()
     serializer_class = CategorySerializer
-    permission_classes = [IsAdmin, AllowAny,] 
+    permission_classes = [IsAdmin, AllowAny, ]
     filter_backends = (filters.SearchFilter,)
     search_fields = ('name',)
 
@@ -64,3 +75,77 @@ class GenreViewSet(GetListCreateDeleteMixin):
     permission_classes = [IsAdmin, AllowAny]
     filter_backends = (filters.SearchFilter,)
     search_fields = ('slug',)
+
+
+@api_view(["POST"])
+@permission_classes([permissions.AllowAny])
+def get_jwt_token(request):
+    serializer = TokenSerializer(data=request.data)
+    serializer.is_valid(raise_exception=True)
+    user = get_object_or_404(
+        User,
+        username=serializer.validated_data["username"]
+    )
+
+    if default_token_generator.check_token(
+        user, serializer.validated_data["confirmation_code"]
+    ):
+        token = AccessToken.for_user(user)
+        return Response({"token": str(token)}, status=status.HTTP_200_OK)
+
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(["POST"])
+@permission_classes([permissions.AllowAny])
+def register(request):
+    serializer = RegisterSerializer(data=request.data)
+    serializer.is_valid(raise_exception=True)
+    serializer.save()
+    user = get_object_or_404(
+        User,
+        username=serializer.validated_data["username"]
+    )
+    confirmation_code = default_token_generator.make_token(user)
+    send_mail(
+        subject="YaMDb registration",
+        message=f"Your confirmation code: {confirmation_code}",
+        from_email=None,
+        recipient=user.email,
+    )
+
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+class UserViewSet(viewsets.ModelViewSet):
+    lookup_field = "username"
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    pagination_class = PageNumberPagination
+    permission_classes = (IsAdmin,)
+
+    @action(
+        methods=[
+            "GET",
+            "PATCH",
+        ],
+        detail=False,
+        url_path="me",
+        permission_classes=permissions.IsAuthenticated,
+        serializer_class=UserEditSerializer,
+    )
+    def users_own_profile(self, request):
+        user = request.user
+        if request.method == "GET":
+            serializer = self.get_serializer(user)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        if request.method == "PATCH":
+            serializer = self.get_serializer(
+                user,
+                data=request.data,
+                partial=True
+            )
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
