+++ api_yamdb/api_yamdb/urls.py
@@ -5,7 +5,6 @@ from django.views.generic import TemplateView
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('api.urls')),
-    path('users/', include('users.urls')),
     path(
         'redoc/',
         TemplateView.as_view(template_name='redoc.html'),
