+++ api_yamdb/titles/models.py
@@ -56,6 +56,6 @@ class GenreTitle(models.Model):
                 name='unique_title_genre'
             )
         ]
-    
+
     def __str__(self):
         return f'{self.title} {self.genre}'
